+++ AndroidHomeApplication/app/build.gradle
@@ -66,9 +66,7 @@ dependencies {
     implementation 'com.squareup.okhttp3:logging-interceptor:4.9.0'
     implementation 'com.squareup.okhttp3:okhttp:4.9.0'
 
-    implementation 'org.jetbrains.kotlinx:kotlinx-serialization-json:1.2.0'
     implementation 'com.jakewharton.retrofit:retrofit2-kotlinx-serialization-converter:0.8.0'
-    implementation 'androidx.paging:paging-runtime-ktx:3.1.0-alpha01'
 
     implementation "androidx.room:room-runtime:2.3.0"
     implementation "androidx.room:room-ktx:2.3.0"
