--- AndroidHomeApplication/app/src/main/java/com/example/androidhomeapplication/data/db/DbProvider.kt
@@ -1,7 +1,0 @@
-package com.example.androidhomeapplication.data.db
-
-import com.example.androidhomeapplication.data.repository.MoviesRepository
-
-interface DbProvider {
-    val db: FilmDatabase
-}
