+++ AndroidHomeApplication/app/src/main/java/com/example/androidhomeapplication/data/db/RoomConverters.kt
@@ -7,9 +7,6 @@ import com.example.androidhomeapplication.data.models.Genre
 import com.squareup.moshi.Moshi
 import com.squareup.moshi.Types
 import com.squareup.moshi.kotlin.reflect.KotlinJsonAdapterFactory
-import kotlinx.serialization.decodeFromString
-import kotlinx.serialization.encodeToString
-import kotlinx.serialization.json.Json
 import java.io.IOException
 
 object RoomConverters {
