+++ AndroidHomeApplication/app/src/main/java/com/example/androidhomeapplication/data/remote/response/MovieResponse.kt
@@ -1,11 +1,10 @@
 package com.example.androidhomeapplication.data.remote.response
 
 import android.os.Parcelable
-import com.example.androidhomeapplication.Utils
+import com.example.androidhomeapplication.utils.Utils
 import com.example.androidhomeapplication.data.db.MovieDbEntity
 import com.example.androidhomeapplication.data.models.Genre
 import com.example.androidhomeapplication.data.models.Movie
-import com.example.androidhomeapplication.data.models.mapToMovieDbEntity
 import com.squareup.moshi.Json
 import kotlinx.android.parcel.Parcelize
