+++ AndroidHomeApplication/app/src/main/java/com/example/androidhomeapplication/utils/DataResult.kt
@@ -1,9 +1,9 @@
-package com.example.androidhomeapplication
+package com.example.androidhomeapplication.utils
 
 
 sealed class DataResult<out T : Any?> {
     class Loading : DataResult<Nothing>()
     class EmptyResult : DataResult<Nothing>()
-    class Success<T>(val value: T) : DataResult<T>()
+    class Success<out T : Any>(val value: T) : DataResult<T>()
     class Error(val error: Throwable) : DataResult<Nothing>()
 }
