+++ AndroidHomeApplication/app/src/main/java/com/example/androidhomeapplication/utils/Utils.kt
@@ -1,4 +1,4 @@
-package com.example.androidhomeapplication
+package com.example.androidhomeapplication.utils
 
 
 object Utils {
